+++ lib/ires/core.rb
@@ -1,13 +1,8 @@
 require 'ffi'
-require 'ires/util'
 
 module Ires
-  module Init
+  module Core
     extend FFI::Library
-    extend Ires::Util
-
-    os = current_os
-    return if os.nil?
 
     # NOTE: ires.so is golang object
     ffi_lib File.expand_path("../../shared/#{os}/ires.so", File.dirname(__FILE__))
